Split functions into different helper modules.
