+++ helpers/misc.py
@@ -1,0 +1,24 @@
+import datetime
+from typing import Union
+
+
+def sort_dict_keys(d: Union[dict, list]) -> Union[dict, list]:
+    """
+    Sort dictionaries and keys.
+    :param d: The object to sort.
+    :return: A sorted list or dictionary.
+    """
+    if type(d) == list:
+        return [sort_dict_keys(e) for e in d]
+    if type(d) != dict:
+        return d
+    return {key: sort_dict_keys(d[key]) for key in sorted(d)}
+
+
+def ts_to_date(timestamp: float) -> str:
+    """
+    Converts timestamp to human readable time.
+    :param timestamp: The epoch timestamp.
+    :return: A human readable time.
+    """
+    return str(datetime.datetime.fromtimestamp(timestamp)).replace(' ', 'T')
